Remove old effects system
