+++ packages/flame/lib/src/effects/component_effect.dart
