+++ packages/flame/lib/src/effects/effect.dart
