+++ packages/flame/lib/src/effects/effect_controller.dart
