+++ packages/flame/lib/src/effects/remove_effect.dart
