+++ packages/flame/lib/src/effects/simple_effect_controller.dart
