+++ packages/flame/lib/src/effects/standard_effect_controller.dart
