+++ packages/flame/test/effects/effect_test.dart
@@ -1,7 +1,7 @@
 import 'package:flame/src/components/component.dart';
-import 'package:flame/src/effects2/effect.dart';
-import 'package:flame/src/effects2/effect_controller.dart';
-import 'package:flame/src/effects2/standard_effect_controller.dart';
+import 'package:flame/src/effects/effect.dart';
+import 'package:flame/src/effects/effect_controller.dart';
+import 'package:flame/src/effects/standard_effect_controller.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
