+++ packages/flame/test/effects/opacity_effect_test.dart
@@ -2,8 +2,8 @@ import 'dart:math';
 
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
-import 'package:flame/src/effects2/opacity_effect.dart';
-import 'package:flame/src/effects2/standard_effect_controller.dart';
+import 'package:flame/src/effects/opacity_effect.dart';
+import 'package:flame/src/effects/standard_effect_controller.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
