+++ packages/flame/test/effects/standard_effect_controller_test.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:flame/src/effects2/standard_effect_controller.dart';
+import 'package:flame/src/effects/standard_effect_controller.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter/animation.dart';
 import 'package:flutter_test/flutter_test.dart';
